requirements.txt
fastapi
paddleocr==2.6.1.3
paddlepaddle==2.6.2
uvicorn
numpy
Pillow 